Cast result to uint64_t to match SYSCALL_ERROR type and eliminate
compiler warning about ternary operator operand signedness mismatch.

Warning fixed:
  kernel/core/syscall.c:727: warning: operand of '?:' changes
  signedness from 'int' to 'long unsigned int' due to unsignedness
  of other operand [-Wsign-compare]

+++ kernel/core/syscall.c
@@ -736,7 +736,7 @@ uint64_t sys_execve(const char *path, const char *argv[], const char *envp[]) {
     int result = elf_load_exec(path, argv, argc);
     
     // If we get here, exec failed
-    return (result < 0) ? SYSCALL_ERROR : result;
+    return (result < 0) ? SYSCALL_ERROR : (uint64_t)result;
 }
 
 /**
